ml-service/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# ML Core
numpy==1.24.3
scipy==1.11.3
scikit-learn==1.3.2
torch==2.1.0

# HuggingFace (for baseline)
transformers==4.35.2

# Data Processing
pandas==2.1.3
joblib==1.3.2

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
httpx==0.25.1

# Optional: For Jupyter notebooks
# jupyter==1.0.0
# matplotlib==3.8.2
# seaborn==0.13.0
